refact: Added proteção nas rotas Carteiras
- Added rotas de Carteiras para Superuser
- Ajustado rotas padrão para buscar o grupo_id do usuario atual
- Ajustado rota path para validar duplicidade
- Ajustado função de duplicidade para suportar o usuario atual e o objeto atual a ser editado

+++ app/routers/carteiras.py
@@ -8,16 +8,21 @@ from sqlalchemy.orm import Session
 from app.schemas.carteiras import *
 from app.utils.text_validator import verificar_duplicidade
 from sqlalchemy import select
+from app.database.models import UsuarioModel
+from app.core.auth import current_user, current_user_superuser
 
 
 roteador = APIRouter(prefix="/carteiras", tags=["Carteiras"])
 
 #--------------------------
-# GET - Todas as Carteiras
-# Rota: GET "/carteiras"
+# GET - Todas as Carteiras - SUPERUSER
+# Rota: GET "/carteiras/admin"
 #--------------------------
-@roteador.get("", response_model=list[CarteiraRead])
-async def todas_carteiras(db: AsyncSession = Depends(get_db)):
+@roteador.get("/admin", response_model=list[CarteiraRead])
+async def todas_carteiras_superuser(
+    db: AsyncSession = Depends(get_db),
+    admin: UsuarioModel = Depends(current_user_superuser)
+):
     
     # Escreve a consulta(execute()), envia ao banco e aguarda a resposta(await).
     carteiras = await db.execute(select(CarteiraModel))
@@ -26,12 +31,35 @@ async def todas_carteiras(db: AsyncSession = Depends(get_db)):
 
 
 #--------------------------
-# POST - Criar Carteira
-# Rota: POST "/carteiras"
+# GET - Todas as Carteiras do usuário atual
+# Rota: GET "/carteiras"
+#--------------------------
+@roteador.get("", response_model=list[CarteiraRead])
+async def todas_carteiras(
+    db: AsyncSession = Depends(get_db),
+    user: UsuarioModel = Depends(current_user)
+):
+    
+    # Escreve a consulta(execute()), envia ao banco e aguarda a resposta(await).
+    carteiras = await db.execute(
+        select(CarteiraModel)
+        .where(CarteiraModel.grupo_familiar_id == user.grupo_familiar_id)
+    )
+    # Faz a extração do resultado para objetos, como é lista usa scalars no plural.
+    return carteiras.scalars().all()
+
+
+#--------------------------
+# POST - Criar Carteira - SUPERUSER
+# Rota: POST "/carteiras/admin"
 #--------------------------
 # status_code é necessário para informar o resultado esperado da requisição
-@roteador.post("", response_model=CarteiraRead, status_code=201)
-async def criar_carteira(payload: CarteiraCreate, db: AsyncSession = Depends(get_db)):
+@roteador.post("/admin", response_model=CarteiraRead, status_code=201)
+async def criar_carteira_superuser(
+    payload: CarteiraCreate,
+    db: AsyncSession = Depends(get_db),
+    admin: UsuarioModel = Depends(current_user_superuser)
+):
     
     # Verifica se tem duplicidade no banco de dados
     await verificar_duplicidade(
@@ -52,18 +80,70 @@ async def criar_carteira(payload: CarteiraCreate, db: AsyncSession = Depends(get
     await db.refresh(nova_carteira)
     return nova_carteira
 
+
 #--------------------------
-# PATCH - Atualizar Carteira
-# Rota: PATCH "/carteiras/[id]"
+# POST - Criar Carteira para usuário atual
+# Rota: POST "/carteiras"
 #--------------------------
-@roteador.patch("/{carteira_id}", response_model=CarteiraRead)
-async def editar_carteira(carteira_id: UUID, payload: CarteiraUpdate, db: AsyncSession = Depends(get_db)):
+# status_code é necessário para informar o resultado esperado da requisição
+@roteador.post("", response_model=CarteiraRead, status_code=201)
+async def criar_carteira(
+    payload: CarteiraCreate,
+    db: AsyncSession = Depends(get_db),
+    user: UsuarioModel = Depends(current_user)
+):
+    if payload.grupo_familiar_id != user.grupo_familiar_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Você não tem permissão para criar uma carteira para este Grupo Familiar")
+    
+    # Verifica se tem duplicidade no banco de dados
+    await verificar_duplicidade(
+        db,
+        model=CarteiraModel,
+        campo="titulo",
+        valor=payload.titulo,
+        mensagem_erro=f"Já existe uma Carteira criada com o título '{payload.titulo}'",
+        cur_user=user
+    )
+
+    # O model_dump pega o objeto payload com as propriedades e desempacota em formato JSON
+    # Os asteriscos (**) serve para entregar ao constructor CarteiraModel as propriedades uma a uma
+    nova_carteira = CarteiraModel(**payload.model_dump())
+    db.add(nova_carteira)
+    await db.commit()
+    # Refresh() é usado para atualizar o objeto do "novo_lancamento" com os dados mais recentes do banco de dados, incluindo o ID gerado automaticamente.
+    # Caso nao queira retornar, deve remover essa linha, remover o response_model e ajustar o return.
+    await db.refresh(nova_carteira)
+    return nova_carteira
+
+
+#--------------------------
+# PATCH - Atualizar Carteira - SUPERUSER
+# Rota: PATCH "/carteiras/admin/[id]"
+#--------------------------
+@roteador.patch("/admin/{carteira_id}", response_model=CarteiraRead)
+async def editar_carteira_superuser(
+    carteira_id: UUID,
+    payload: CarteiraUpdate,
+    db: AsyncSession = Depends(get_db),
+    admin: UsuarioModel = Depends(current_user_superuser)
+):
     # Busca o objeto no Banco pesquisando pelo id passado
     obj_alvo = await db.get(CarteiraModel, carteira_id)
 
     # Trata o erro caso não encontre o objeto
     if not obj_alvo:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Carteira não encontrada")
+    
+    # Valida duplicidade, ignorando o objeto atual
+    if payload.titulo:
+        await verificar_duplicidade(
+            db,
+            model=CarteiraModel,
+            campo="titulo",
+            valor=payload.titulo,
+            mensagem_erro=f"Já existe uma Carteira criada com o título '{payload.titulo}'",
+            id_ignorar=carteira_id
+        )
     
     # O model_dump desmonta o objeto em JSON
     # O CarteiraUpdate monta um Dic com a propriedades definidas, se usuario mandar apenas uma o exclude_unset remove a outra
@@ -81,16 +161,92 @@ async def editar_carteira(carteira_id: UUID, payload: CarteiraUpdate, db: AsyncS
 
 
 #--------------------------
-# DELETE - Remover Carteira
-# Rota: DELETE "/carteiras/[id]"
+# PATCH - Atualizar Carteira do usuário atual
+# Rota: PATCH "/carteiras/[id]"
 #--------------------------
-@roteador.delete("/{carteira_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def remover_carteira(carteira_id: UUID, db: AsyncSession = Depends(get_db)):
+@roteador.patch("/{carteira_id}", response_model=CarteiraRead)
+async def editar_carteira(
+    carteira_id: UUID,
+    payload: CarteiraUpdate,
+    db: AsyncSession = Depends(get_db),
+    user: UsuarioModel = Depends(current_user)
+):
+    # Busca o objeto no Banco pesquisando pelo id passado
+    obj_alvo = await db.get(CarteiraModel, carteira_id)
+
+    # Trata o erro caso não encontre o objeto
+    if not obj_alvo:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Carteira não encontrada")
+    
+    if obj_alvo.grupo_familiar_id != user.grupo_familiar_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Você não tem permissão para editar esta carteira")
+    
+    # Valida duplicidade, ignorando o objeto atual
+    if payload.titulo:
+        await verificar_duplicidade(
+            db,
+            model=CarteiraModel,
+            campo="titulo",
+            valor=payload.titulo,
+            mensagem_erro=f"Já existe uma Carteira criada com o título '{payload.titulo}'",
+            cur_user=user,
+            id_ignorar=carteira_id
+        )
+    
+    # O model_dump desmonta o objeto em JSON
+    # O CarteiraUpdate monta um Dic com a propriedades definidas, se usuario mandar apenas uma o exclude_unset remove a outra
+    novos_dados = payload.model_dump(exclude_unset=True)
+
+    # Variavel campo e valor recebem respectivamente do Dic enviado pelo usuario
+    # setattr pega o objeto do banco e altera o valor com base na propriedade
+    for campo, valor in novos_dados.items():
+        setattr(obj_alvo, campo, valor)
+    
+    db.add(obj_alvo)
+    await db.commit()
+    await db.refresh(obj_alvo)
+    return obj_alvo
+
+
+#--------------------------
+# DELETE - Remover Carteira - SUPERUSER
+# Rota: DELETE "/carteiras/admin/[id]"
+#--------------------------
+@roteador.delete("/admin/{carteira_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def remover_carteira_superuser(
+    carteira_id: UUID,
+    db: AsyncSession = Depends(get_db),
+    admin: UsuarioModel = Depends(current_user_superuser)
+):
 
     obj_alvo = await db.get(CarteiraModel, carteira_id)
 
     if not obj_alvo:
-        raise HTTPException(status_code=status.HTTP_204_NO_CONTENT, detail="Carteira não encontrada")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Carteira não encontrada")
+
+    await db.delete(obj_alvo)
+    await db.commit()
+    return None
+
+
+#--------------------------
+# DELETE - Remover Carteira do usuário atual
+# Rota: DELETE "/carteiras/[id]"
+#--------------------------
+@roteador.delete("/{carteira_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def remover_carteira(
+    carteira_id: UUID,
+    db: AsyncSession = Depends(get_db),
+    user: UsuarioModel = Depends(current_user)
+):
+
+    obj_alvo = await db.get(CarteiraModel, carteira_id)
+
+    if not obj_alvo:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Carteira não encontrada")
+    
+    if obj_alvo.grupo_familiar_id != user.grupo_familiar_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Você não tem permissão para excluir esta carteira")
 
     await db.delete(obj_alvo)
     await db.commit()
